+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### RustZX v0.14
+- **[Refactoring]** Moved z80 emulation to the separate crate (`rustzx-z80`)
+- **[Feature]** Added support for `*.scr` screenshot format loading
+
 ### RustZX v0.13
 - **[Feature]** Support for ZX Spectrum 128K SNA variant
 - **[Feature]** Quick save/load (`F1`/`F2` keys) via SNA